requirements.txt
numpy
matplotlib
Pillow
comtypes
